+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Unreleased changes are in the `master` branch.
 
 ## Fixed
 
-- `/utils/txs/evaluate` example
+- `/utils/txs/evaluate` example and Ogmios API Reference URL
 
 ## [0.1.52] 2023-01-02